cleaning matrix2D header
